combined_requirements.txt
requests>=2.31.0
pydantic>=2.4.2
minio>=7.1.17
python-dotenv>=1.0.0
Flask>=2.0.0
urllib3>=2.0.0